Struct leo_compiler::compiler::SymbolTable

pub struct SymbolTable {
    pub(crate) parent: Option<Box<SymbolTable, Global>>,
    pub functions: IndexMap<Symbol, FunctionSymbol, RandomState>,
    pub circuits: IndexMap<Symbol, Circuit, RandomState>,
    pub(crate) variables: IndexMap<Symbol, VariableSymbol, RandomState>,
    pub(crate) scope_index: usize,
    pub(crate) scopes: Vec<RefCell<SymbolTable>, Global>,
}

Fields
parent: Option<Box<SymbolTable, Global>>

The parent scope if it exists. For example, the parent scope of a then-block is the scope containing the associated ConditionalStatement.

functions: IndexMap<Symbol, FunctionSymbol, RandomState>

Functions represents the name of each function mapped to the AST’s function definition. This field is populated at a first pass.

circuits: IndexMap<Symbol, Circuit, RandomState>

Maps circuit names to circuit definitions. This field is populated at a first pass.

variables: IndexMap<Symbol, VariableSymbol, RandomState>

The variables defined in a scope. This field is populated as necessary.

scope_index: usize

The index of the current scope.

scopes: Vec<RefCell<SymbolTable>, Global>

The sub-scopes of this scope.

Implementations

impl SymbolTable

pub fn check_shadowing(
&self,
symbol: Symbol,
span: Span
) -> Result<(), LeoError>

Recursively checks if the symbol table contains an entry for the given symbol. Leo does not allow any variable shadowing or overlap between different symbols.

pub fn scope_index(&mut self) -> usize

Returns the current scope index. Increments the scope index.

pub fn insert_fn(
&mut self,
symbol: Symbol,
insert: &Function
) -> Result<(), LeoError>

Inserts a function into the symbol table.

pub fn insert_circuit(
&mut self,
symbol: Symbol,
insert: &Circuit
) -> Result<(), LeoError>

Inserts a circuit into the symbol table.

pub fn insert_variable(
&mut self,
symbol: Symbol,
insert: VariableSymbol
) -> Result<(), LeoError>

Inserts a variable into the symbol table.

pub fn insert_block(&mut self) -> usize

Creates a new scope for the block and stores it in the symbol table.

pub fn lookup_fn_symbol(&self, symbol: Symbol) -> Option<&FunctionSymbol>

Attempts to lookup a function in the symbol table.

pub fn lookup_circuit(&self, symbol: Symbol) -> Option<&Circuit>

Attempts to lookup a circuit in the symbol table.

pub fn lookup_variable(&self, symbol: Symbol) -> Option<&VariableSymbol>

Attempts to lookup a variable in the symbol table.

pub fn variable_in_local_scope(&self, symbol: Symbol) -> bool

Returns true if the variable exists in the local scope

pub fn variable_in_parent_scope(&self, symbol: Symbol) -> bool

Returns true if the variable exists in any parent scope

pub fn lookup_variable_mut(
&mut self,
symbol: Symbol
) -> Option<&mut VariableSymbol>

Returns a mutable reference to the VariableSymbol if it exists in the symbol table.

pub fn lookup_fn_scope(&self, symbol: Symbol) -> Option<&RefCell<SymbolTable>>

Returns the scope associated with the function symbol, if it exists in the symbol table.

pub fn lookup_scope_by_index(
&self,
index: usize
) -> Option<&RefCell<SymbolTable>>

Returns the scope associated with index, if it exists in the symbol table.