Struct leo_compiler::program::program::ConstrainedProgram

pub struct ConstrainedProgram<'a, F: PrimeField, G: GroupType<F>> {
    pub asg: Program<'a>,
    identifiers: IndexMap<u32, ConstrainedValue<'a, F, G>>,
}

Fields
asg: Program<'a>identifiers: IndexMap<u32, ConstrainedValue<'a, F, G>>
Implementations

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn evaluate_console_assert<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
indicator: &Boolean,
expression: &'a Expression<'a>,
span: &Span
) -> Result<()>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn evaluate_console_function_call<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
indicator: &Boolean,
console: &ConsoleStatement<'a>
) -> Result<()>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn format<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
args: &ConsoleArgs<'a>
) -> Result<String>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn store_definition(
&mut self,
variable: &Variable<'_>,
value: ConstrainedValue<'a, F, G>
)

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_array<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
array: &[(Cell<&'a Expression<'a>>, bool)],
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

Enforce array expressions

pub fn enforce_array_initializer<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
element_expression: &'a Expression<'a>,
actual_size: usize
) -> Result<ConstrainedValue<'a, F, G>>

Returns an array value from an array initializer expression.

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn array_bounds_check<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
index_resolved: &Integer,
array_len: u32,
span: &Span
) -> Result<()>

pub fn enforce_array_access<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
array: &'a Expression<'a>,
index: &'a Expression<'a>,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

pub fn enforce_array_range_access<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
array: &'a Expression<'a>,
left: Option<&'a Expression<'a>>,
right: Option<&'a Expression<'a>>,
length: usize,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub(crate) fn enforce_index<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
index: &'a Expression<'a>,
span: &Span
) -> Result<Integer>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_binary_expression<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
left: &'a Expression<'a>,
right: &'a Expression<'a>
) -> Result<(ConstrainedValue<'a, F, G>, ConstrainedValue<'a, F, G>)>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_circuit_access<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
expr: &CircuitAccessExpression<'a>
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_circuit<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
expr: &CircuitInitExpression<'a>,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_conditional_expression<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
conditional: &'a Expression<'a>,
first: &'a Expression<'a>,
second: &'a Expression<'a>,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

Enforce ternary conditional expression

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub(crate) fn enforce_const_value<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
value: &'a ConstValue<'a>,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

pub(crate) fn enforce_expression<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
expression: &'a Expression<'a>
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_core_circuit_call_expression<CS: ConstraintSystem<F>, C: CoreCircuit<'a, F, G>>(
&mut self,
cs: &mut CS,
core_circuit: &C,
function: &'a Function<'a>,
target: Option<&'a Expression<'a>>,
arguments: &[Cell<&'a Expression<'a>>],
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

Call a default core circuit function with arguments

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_function_call_expression<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
function: &'a Function<'a>,
target: Option<&'a Expression<'a>>,
arguments: &[Cell<&'a Expression<'a>>],
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn evaluate_ref(
&mut self,
variable_ref: &VariableRef<'_>
) -> Result<ConstrainedValue<'a, F, G>>

Enforce a variable expression by getting the resolved value

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_lengthof<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
lengthof: &'a LengthOfExpression<'a>,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

Enforce array expressions

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_tuple_access<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
tuple: &'a Expression<'a>,
index: usize,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_tuple<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
tuple: &[Cell<&'a Expression<'a>>]
) -> Result<ConstrainedValue<'a, F, G>>

Enforce tuple expressions

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn allocate_array<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
name: &str,
array_type: &Type<'_>,
array_len: usize,
input_value: Option<InputValue>,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn allocate_main_function_input<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
type_: &Type<'_>,
name: &str,
input_option: Option<InputValue>,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

Process constant inputs and return ConstrainedValue with constant value in it.

pub fn constant_main_function_input<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
type_: &Type<'_>,
name: &str,
input_option: Option<InputValue>,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn allocate_input_keyword<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
span: &Span,
expected_type: &'a Circuit<'a>,
input: &Input
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn allocate_input_section<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
identifier: Identifier,
expected_type: &'a Circuit<'a>,
section: IndexMap<Parameter, Option<InputValue>>
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn allocate_tuple<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
name: &str,
types: &[Type<'_>],
input_value: Option<InputValue>,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub(crate) fn enforce_function<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
function: &'a Function<'a>,
target: Option<&'a Expression<'a>>,
arguments: &[Cell<&'a Expression<'a>>]
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_main_function<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
function: &'a Function<'a>,
input: &Input
) -> Result<Output>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn conditionally_select_result<CS: ConstraintSystem<F>>(
cs: &mut CS,
expected_return: &Type<'a>,
results: Vec<(Boolean, ConstrainedValue<'a, F, G>)>,
span: &Span
) -> Result<ConstrainedValue<'a, F, G>>

Returns a conditionally selected result from the given possible function returns and given function return type.

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

fn prepare_mut_access(
out: &mut Vec<AssignAccess<'a>>,
expr: &'a Expression<'a>
) -> Result<Option<&'a Variable<'a>>>

pub fn resolve_mut_ref<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
assignee: &'a Expression<'a>,
target_value: ConstrainedValue<'a, F, G>,
indicator: &Boolean
) -> Result<bool>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn new(asg: Program<'a>) -> Self

pub(crate) fn store(&mut self, id: u32, value: ConstrainedValue<'a, F, G>)

pub(crate) fn get(&self, id: u32) -> Option<&ConstrainedValue<'a, F, G>>

pub(crate) fn get_mut(
&mut self,
id: u32
) -> Option<&mut ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_assign_statement<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
indicator: &Boolean,
statement: &AssignStatement<'a>
) -> Result<()>

pub(super) fn enforce_assign_operation<CS: ConstraintSystem<F>>(
cs: &mut CS,
condition: &Boolean,
scope: String,
operation: &AssignOperation,
target: &mut ConstrainedValue<'a, F, G>,
new_value: ConstrainedValue<'a, F, G>,
span: &Span
) -> Result<()>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub(super) fn resolve_target_access_array_index<'b, CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
context: ResolverContext<'a, 'b, F, G>,
index: &'a Expression<'a>
) -> Result<()>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub(super) fn resolve_target_access_array_range<'b, CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
context: ResolverContext<'a, 'b, F, G>,
start: Option<&'a Expression<'a>>,
stop: Option<&'a Expression<'a>>
) -> Result<()>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub(super) fn resolve_target_access_member<'b, CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
context: ResolverContext<'a, 'b, F, G>,
name: &Identifier
) -> Result<()>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub(super) fn resolve_target_access_tuple<'b, CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
context: ResolverContext<'a, 'b, F, G>,
index: usize
) -> Result<()>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

fn enforce_assign_context<'b, CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
context: &ResolverContext<'a, 'b, F, G>,
target: &mut ConstrainedValue<'a, F, G>
) -> Result<()>

fn resolve_target_access<'b, CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
context: ResolverContext<'a, 'b, F, G>
) -> Result<()>

pub fn resolve_assign<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
assignee: &AssignStatement<'a>,
target_value: ConstrainedValue<'a, F, G>,
indicator: &Boolean
) -> Result<()>

pub(crate) fn check_range_index(
start_index: usize,
stop_index: usize,
len: usize,
span: &Span
) -> Result<()>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn evaluate_block<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
indicator: &Boolean,
block: &BlockStatement<'a>
) -> StatementResult<Vec<IndicatorAndConstrainedValue<'a, F, G>>>

Evaluates a branch of one or more statements and returns a result in the given scope.

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_conditional_statement<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
indicator: &Boolean,
statement: &ConditionalStatement<'a>
) -> StatementResult<Vec<IndicatorAndConstrainedValue<'a, F, G>>>

Enforces a conditional statement with one or more branches. Due to R1CS constraints, we must evaluate every branch to properly construct the circuit. At program execution, we will pass an indicator bit down to all child statements within each branch. The indicator bit will select that branch while keeping the constraint system satisfied.

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

fn enforce_multiple_definition(
&mut self,
variable_names: &[&'a Variable<'a>],
values: Vec<ConstrainedValue<'a, F, G>>,
span: &Span
) -> Result<()>

pub fn enforce_definition_statement<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
statement: &DefinitionStatement<'a>
) -> Result<()>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_iteration_statement<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
indicator: &Boolean,
statement: &IterationStatement<'a>
) -> StatementResult<Vec<IndicatorAndConstrainedValue<'a, F, G>>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_return_statement<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
statement: &ReturnStatement<'a>
) -> Result<ConstrainedValue<'a, F, G>>

impl<'a, F: PrimeField, G: GroupType<F>> ConstrainedProgram<'a, F, G>

pub fn enforce_statement<CS: ConstraintSystem<F>>(
&mut self,
cs: &mut CS,
indicator: &Boolean,
statement: &'a Statement<'a>
) -> StatementResult<Vec<IndicatorAndConstrainedValue<'a, F, G>>>

Enforce a program statement. Returns a Vector of (indicator, value) tuples. Each evaluated statement may execute of one or more statements that may return early. To indicate which of these return values to take we conditionally select the value according to the indicator bit that evaluates to true.