leo_compiler

Module run_with_ledger

Structs

Case

A particular case to run.

CaseOutcome

All details about the result of a case that was run.

Config

Programs and configuration to run.

Program

A program to deploy to the ledger.

Enums

Status

The status of a case that was run.

Functions

run_with_ledger

Run the functions indicated by cases from the programs in config.

Type Aliases

CurrentNetwork ๐