Module leo_compiler::statement::statement

Enforces a statement in a compiled Leo program.

Functions

get_indicator_value
Unwraps the indicator boolean gadget value or false if None. This method is used by logging methods only. We can directly get the boolean value of the indicator since we are not enforcing any constraints.

Type Definitions

IndicatorAndConstrainedValue

StatementResult