leo_compiler

Struct Ast

pub struct Ast {
    pub ast: Program,
}

The abstract syntax tree (AST) for a Leo program.

The Ast type represents a Leo program as a series of recursive data types. These data types form a tree that begins from a Program type root.

Fields
ast: Program
Implementations

impl Ast

pub fn new(program: Program) -> Ast

Creates a new AST from a given program tree.

pub fn as_repr(&self) -> &Program

Returns a reference to the inner program AST representation.

pub fn into_repr(self) -> Program

pub fn to_json_string(&self) -> Result<String, LeoError>

Serializes the ast into a JSON string.

pub fn to_json_value(&self) -> Result<Value, LeoError>

pub fn to_json_file( &self, path: PathBuf, file_name: &str, ) -> Result<(), LeoError>

Serializes the ast into a JSON file.

pub fn to_json_file_without_keys( &self, path: PathBuf, file_name: &str, excluded_keys: &[&str], ) -> Result<(), LeoError>

Serializes the ast into a JSON value and removes keys from object mappings before writing to a file.

pub fn from_json_string(json: &str) -> Result<Ast, LeoError>

Deserializes the JSON string into a ast.

pub fn from_json_file(path: PathBuf) -> Result<Ast, LeoError>

Deserializes the JSON string into a ast from a file.