Struct leo_compiler::SymbolTable

pub struct SymbolTable {
    pub(crate) parent: Option<Box<SymbolTable>>,
    pub functions: IndexMap<Location, FunctionSymbol>,
    pub structs: IndexMap<Location, Composite>,
    pub(crate) variables: IndexMap<Location, VariableSymbol>,
    pub(crate) scope_index: usize,
    pub(crate) scopes: Vec<RefCell<SymbolTable>>,
}

Fields
parent: Option<Box<SymbolTable>>functions: IndexMap<Location, FunctionSymbol>

Maps parent program name and function name to the AST’s function definition. This field is populated at a first pass.

structs: IndexMap<Location, Composite>

Maps parent program name and composite name to composite definitions. This field is populated at a first pass.

variables: IndexMap<Location, VariableSymbol>scope_index: usizescopes: Vec<RefCell<SymbolTable>>
Implementations

impl SymbolTable

pub fn check_shadowing( &self, location: &Location, is_struct: bool, span: Span, ) -> Result<(), LeoError>

Recursively checks if the symbol table contains an entry for the given symbol. Leo does not allow any variable shadowing or overlap between different symbols.

pub fn scope_index(&mut self) -> usize

Returns the current scope index. Increments the scope index.

pub fn insert_fn( &mut self, location: Location, insert: &Function, ) -> Result<(), LeoError>

Inserts a function into the symbol table.

pub fn insert_struct( &mut self, location: Location, insert: &Composite, ) -> Result<(), LeoError>

Inserts a struct into the symbol table.

pub fn attach_finalize( &mut self, caller: Location, callee: Location, ) -> Result<(), LeoError>

Attach a finalize to a function.

pub fn insert_variable( &mut self, location: Location, insert: VariableSymbol, ) -> Result<(), LeoError>

Inserts a variable into the symbol table.

pub fn insert_futures( &mut self, program: Symbol, function: Symbol, futures: Vec<Location>, ) -> Result<(), LeoError>

Inserts futures into the function definition.

pub fn remove_variable_from_current_scope(&mut self, location: Location)

Removes a variable from the symbol table.

pub fn insert_block(&mut self) -> usize

Creates a new scope for the block and stores it in the symbol table.

pub fn lookup_fn_symbol(&self, location: Location) -> Option<&FunctionSymbol>

Attempts to lookup a function in the symbol table.

pub fn lookup_struct( &self, location: Location, main_program: Option<Symbol>, ) -> Option<&Composite>

Attempts to lookup a struct in the symbol table.

pub fn lookup_variable(&self, location: Location) -> Option<&VariableSymbol>

Attempts to lookup a variable in the symbol table.

pub fn lookup_variable_in_current_scope( &self, location: Location, ) -> Option<&VariableSymbol>

Attempts to lookup a variable in the current scope.

pub fn lookup_scope_by_index( &self, index: usize, ) -> Option<&RefCell<SymbolTable>>

Returns the scope associated with index, if it exists in the symbol table.

pub fn to_json_string(&self) -> Result<String, LeoError>

Serializes the symbol table into a JSON string.

pub fn to_json_value(&self) -> Result<Value, LeoError>

Converts the symbol table into a JSON value

pub fn to_json_file( &self, path: PathBuf, file_name: &str, ) -> Result<(), LeoError>

pub fn to_json_file_without_keys( &self, path: PathBuf, file_name: &str, excluded_keys: &[&str], ) -> Result<(), LeoError>

Serializes the symbol table into a JSON value and removes keys from object mappings before writing to a file.

pub fn from_json_string(json: &str) -> Result<SymbolTable, LeoError>

Deserializes the JSON string into a symbol table.

pub fn from_json_file(path: PathBuf) -> Result<SymbolTable, LeoError>

Deserializes the JSON string into a symbol table from a file.