Trait leo_compiler::value::integer::macros::IntegerTrait

pub trait IntegerTrait: Debug + Clone {
    type IntegerType;
    type UnsignedGadget;
    type UnsignedIntegerType;

    const SIZE: usize;

    fn constant(value: Self::IntegerType) -> Self;

    fn one() -> Self;

    fn zero() -> Self;

    fn new(bits: Vec<Boolean, Global>, value: Option<Self::IntegerType>) -> Self;

    fn is_constant(&self) -> bool;

    fn to_bits_le(&self) -> Vec<Boolean, Global>;

    fn from_bits_le(bits: &[Boolean]) -> Self;

    fn get_value(&self) -> Option<String>;

    fn result_is_constant(first: &Self, second: &Self) -> bool { ... }
}

The interface for a singed or unsigned integer gadget.

Associated Types

type IntegerType

type UnsignedGadget

type UnsignedIntegerType

Associated Constants

const SIZE: usize

Required methods

fn constant(value: Self::IntegerType) -> Self

fn one() -> Self

fn zero() -> Self

fn new(bits: Vec<Boolean, Global>, value: Option<Self::IntegerType>) -> Self

fn is_constant(&self) -> bool

Returns true if all bits in this Int are constant

fn to_bits_le(&self) -> Vec<Boolean, Global>

fn from_bits_le(bits: &[Boolean]) -> Self

fn get_value(&self) -> Option<String>

Provided methods

fn result_is_constant(first: &Self, second: &Self) -> bool

Returns true if both Int objects have constant bits