Struct leo_compiler::SymbolTable

pub struct SymbolTable {
    pub(crate) parent: Option<Box<SymbolTable, Global>>,
    pub functions: IndexMap<Symbol, FunctionSymbol, RandomState>,
    pub structs: IndexMap<Symbol, Struct, RandomState>,
    pub(crate) variables: IndexMap<Symbol, VariableSymbol, RandomState>,
    pub(crate) scope_index: usize,
    pub(crate) scopes: Vec<RefCell<SymbolTable>, Global>,
}

Fields
parent: Option<Box<SymbolTable, Global>>functions: IndexMap<Symbol, FunctionSymbol, RandomState>

Functions represents the name of each function mapped to the AST’s function definition. This field is populated at a first pass.

structs: IndexMap<Symbol, Struct, RandomState>

Maps struct names to struct definitions. This field is populated at a first pass.

variables: IndexMap<Symbol, VariableSymbol, RandomState>scope_index: usizescopes: Vec<RefCell<SymbolTable>, Global>
Implementations

impl SymbolTable

pub fn check_shadowing( &self, symbol: Symbol, span: Span ) -> Result<(), LeoError>

Recursively checks if the symbol table contains an entry for the given symbol. Leo does not allow any variable shadowing or overlap between different symbols.

pub fn scope_index(&mut self) -> usize

Returns the current scope index. Increments the scope index.

pub fn insert_fn( &mut self, symbol: Symbol, insert: &Function ) -> Result<(), LeoError>

Inserts a function into the symbol table.

pub fn insert_struct( &mut self, symbol: Symbol, insert: &Struct ) -> Result<(), LeoError>

Inserts a struct into the symbol table.

pub fn insert_variable( &mut self, symbol: Symbol, insert: VariableSymbol ) -> Result<(), LeoError>

Inserts a variable into the symbol table.

pub fn remove_variable_from_current_scope(&mut self, symbol: Symbol)

Removes a variable from the symbol table.

pub fn insert_block(&mut self) -> usize

Creates a new scope for the block and stores it in the symbol table.

pub fn lookup_fn_symbol(&self, symbol: Symbol) -> Option<&FunctionSymbol>

Attempts to lookup a function in the symbol table.

pub fn lookup_struct(&self, symbol: Symbol) -> Option<&Struct>

Attempts to lookup a struct in the symbol table.

pub fn lookup_variable(&self, symbol: Symbol) -> Option<&VariableSymbol>

Attempts to lookup a variable in the symbol table.

pub fn variable_in_local_scope(&self, symbol: Symbol) -> bool

Returns true if the variable exists in the local scope

pub fn variable_in_parent_scope(&self, symbol: Symbol) -> bool

Returns true if the variable exists in any parent scope

pub fn lookup_variable_mut( &mut self, symbol: Symbol ) -> Option<&mut VariableSymbol>

Returns a mutable reference to the VariableSymbol if it exists in the symbol table.

pub fn lookup_fn_scope(&self, symbol: Symbol) -> Option<&RefCell<SymbolTable>>

Returns the scope associated with the function symbol, if it exists in the symbol table.

pub fn lookup_scope_by_index( &self, index: usize ) -> Option<&RefCell<SymbolTable>>

Returns the scope associated with index, if it exists in the symbol table.

pub fn to_json_string(&self) -> Result<String, LeoError>

Serializes the symbol table into a JSON string.

pub fn to_json_value(&self) -> Result<Value, LeoError>

Converts the symbol table into a JSON value

pub fn to_json_file( &self, path: PathBuf, file_name: &str ) -> Result<(), LeoError>

pub fn to_json_file_without_keys( &self, path: PathBuf, file_name: &str, excluded_keys: &[&str] ) -> Result<(), LeoError>

Serializes the symbol table into a JSON value and removes keys from object mappings before writing to a file.

pub fn from_json_string(json: &str) -> Result<SymbolTable, LeoError>

Deserializes the JSON string into a symbol table.

pub fn from_json_file(path: PathBuf) -> Result<SymbolTable, LeoError>

Deserializes the JSON string into a symbol table from a file.